Get Fit Spectra

Given a fitted peak object and the FTIR source of the fit, produce a resultant `absorbance` column. Typically used to plot fitted peaks or calculate residuals.

Étant donné un objet pic IRTF ajusté et la source IRTF de l'ajustement, produire une colonne `absorbance` résultante. Généralement utilisé pour tracer les pics ajustés ou calculer les résidus.

Usage

.get_fit_spectra(ftir, fitted_peaks, peak = NULL, call = rlang::caller_env())

Arguments

ftir

A data.frame in long format with a single FTIR spectra in columns `sample_id`, `wavenumber`, and `absorbance`.

Un data.frame au format long avec un seul spectre IRTF dans les colonnes `sample_id`, `wavenumber`, et `absorbance`.

fitted_peaks

An object from [fit_peaks()]. Should match the provided sample in `ftir`.

Un objet de [fit_peaks()]. Doit correspondre à l'échantillon fourni dans `ftir`.

peak

A peak index if getting single peak spectra, else returns the sum of all fitted peaks.

Un index des pics si l'on obtient des spectres à un seul pic, sinon la somme de tous les pics ajustés est renvoyée.

Value

The calculated absorbance intensities as numeric vector of the same length as the FTIR spectra.

Les intensités d'absorption calculées sous forme de tableau numérique de même longueur que les spectres IRTF.

References

Matsumura, T., Nagamura, N., Akaho, S., Nagata, K., & Ando, Y. (2019) "Spectrum adapted expectation-maximization algorithm for high-throughput peak shift analysis". Science and technology of advanced materials, 20(1), pp 733-745. doi:10.1080/14686996.2019.1620123 Matsumura, T., Nagamura, N., Akaho, S., Nagata, K., & Ando, Y. (2021) "Spectrum adapted expectation-conditional maximization algorithm for extending high–throughput peak separation method in XPS analysis". Science and Technology of Advanced Materials: Methods, 1(1), pp 45-55. doi:10.1080/27660400.2021.1899449