Recalculate Baseline

It may be desired to shift the baseline signal (0 for absorbance or 100 for transmittance) to aid in plotting the spectra. This can be done for all samples or a subset, using the same shift for all adjusted samples or calculated individually.

Recalculate or shift to baseline/max transmittance can be done following one of a few methods:

To shift baseline based on the value at a given wavenumber: recalculate_baseline(ftir, wavenumber_range = [numeric], method = 'point')

To shift baseline based on the average value across a provided wavenumber range: recalculate_baseline(ftir, wavenumber_range = c([numeric], [numeric]), method = 'average')

To shift baseline based on the value at the single lowest point of absorbance (or highest point of transmittance) across the whole spectra recalculate_baseline(ftir, method = 'minimum')

To shift baseline based on the value at the single lowest point of absorbance (or highest point of transmittance) in a given range recalculate_baseline(ftir, wavenumber_range = c([numeric], [numeric]), method = 'minimum')

To perform the exact same baseline adjustment on all samples, specify individually = FALSE. To adjust with a unique determination for each sample, specify individualy = TRUE.

Il peut être souhaitable de décaler le signal de la ligne de base (0 pour l'absorbance ou 100 pour la transmittance) pour faciliter le tracé des spectres. Cela peut être fait pour tous les échantillons ou un sous-ensemble, en utilisant le même décalage pour tous les échantillons ajustés ou calculés individuellement.

Le recalcul ou le décalage de la ligne de base/transmittance maximale peut être effectué en suivant l'une des méthodes suivantes :

Pour décaler la ligne de base en fonction de la valeur à un nombre d'ondes donné : recalculate_baseline(ftir, wavenumber_range = [numeric], method = 'point')

Pour décaler la ligne de base en fonction de la valeur moyenne sur un nombre d'ondes donné : #' recalculate_baseline(ftir) = [numerique], method = 'point') recalculate_baseline(ftir, wavenumber_range = c(numeric, numeric), method = 'average')`

Pour décaler la ligne de base en fonction de la valeur du point d'absorbance le plus bas (ou du point de transmittance le plus élevé) sur l'ensemble des spectres. recalculate_baseline(ftir, method = 'minimum')

Décaler la ligne de base en fonction de la valeur du point d'absorbance le plus bas (ou du point de transmittance le plus élevé) dans une gamme donnée. recalculate_baseline(ftir, wavenumber_range = c([numeric], [numeric]), method = 'minimum')

Pour effectuer exactement le même ajustement de la ligne de base sur tous les échantillons, spécifiez individually = FALSE. Pour ajuster avec une détermination unique pour chaque échantillon, spécifiez individualy = TRUE.

Usage

recalculate_baseline(
  ftir,
  sample_ids = NA,
  wavenumber_range = NA,
  method = "average",
  individually = TRUE
)

Arguments

ftir

A data.frame in long format with columns `sample_id`, `wavenumber`, and `absorbance`/`transmittance`. Un data.frame au format long avec les colonnes `sample_id`, `wavenumber`, et `absorbance`/`transmittance`.

sample_ids

A vector of sample IDs to be adjusted. All sample IDs must be present in the ftir data.frame. If adjusting all spectra, provide NA or NULL. Unlisted sample_id from ftir will be left alone.

Un vecteur d'ID d'échantillons à ajuster Tous les ID d'échantillons doivent être présents dans la base de données ftir data.frame. Si l'ajustement concerne tous les spectres, fournir NA ou NULL. Les sample_id non listés de ftir seront laissés seuls.

wavenumber_range

If specifying a single point wavenumber; a single numeric value. If specifying a wavenumber range, then a vector of two numeric values.

Si l'on spécifie un nombre d'ondes ponctuel, une seule valeur numérique. Si l'on spécifie un nombre d'ondes, alors un vecteur de deux valeurs numériques.

method

One of three values:

If adjusting by the value from a specific wavenumber, provide "point",

If adjusting by the average from a range, provide "average".

If adjusting by the minimum (for absorbance) or maximum (for transmittance) from a range or spectra, provide "minimum" or "maximum", the appropriate transformation will be performed based on spectra type.

Une des trois valeurs :

Si l'ajustement se fait par la valeur d'un nombre d'ondes spécifique, fournir "point",

Si l'ajustement se fait par la moyenne d'une gamme, fournir "average".

Si l'ajustement se fait par le minimum (pour l'absorbance) ou le maximum (pour la transmittance) d'une gamme ou de spectres, indiquez "minimum" ou "maximum", la transformation appropriée sera effectuée en fonction du type de spectre.

individually

If adjusting all samples by the same amount, specify TRUE, else specify FALSE for unique adjustments. When TRUE, the smallest absolute individual sample adjustment to achieve baseline will be applied to all named samples.

Si vous ajustez tous les échantillons de la même manière, spécifiez TRUE, sinon spécifiez FALSE pour des ajustements uniques. Si TRUE, le plus petit ajustement absolu d'un échantillon individuel pour atteindre la ligne de base sera appliqué à tous les échantillons nommés.

Value

A data.frame containing the adjusted FTIR spectra.

Un data.frame contenant les spectres IRTF ajustee.

Examples

# Adjust the biodiesel spectra to minimum for each sample
recalculate_baseline(biodiesel, method = "minimum", individually = TRUE)
#> PlotFTIR data:
#>   Spectral range: 700.7395 - 3999.434 cm⁻¹ 
#>   Resolution: variable
#>   Intensity type: absorbance 
#>   Number of samples: 11 
#>   Sample IDs: biodiesel_0, biodiesel_0_25, biodiesel_0_50, biodiesel_1_0, biodiesel_2_5 ...