Save FTIR Plot

Save FTIR plot object to file. Uses [ggplot2::ggsave()] to save to disk. Specify a filename ending with `.svg` for vector graphics, if requeste by a journal.

Enregistrer l'objet de tracé IRTF dans un fichier. Utilise [ggplot2::ggsave()] pour enregistrer sur le disque. Spécifier un nom de fichier se terminant par `.svg` pour les graphiques vectoriels, si un journal le demande.

Usage

save_plot(ftir_spectra_plot, filename, ...)

Arguments

ftir_spectra_plot

A plot generated by [plot_ftir()] or [plot_ftir_stacked()].

Un tracé généré par [plot_ftir()] ou [plot_ftir_stacked()].

filename

Name and directory of the file you wish to create. If it includes a extension the function will produce a file of that type. Options for filetypes include "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (on windows only).

Nom et répertoire du fichier que vous souhaitez créer. S'il contient une extension, la fonction produira un fichier de ce type. Les options pour les types de fichiers incluent "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" ou "wmf" (sur Windows uniquement).

...

Additional arguements to pass to [ggplot2::ggsave()].

Arguments supplémentaires à passer à [ggplot2::ggsave()].

Value

invisible `TRUE`

Examples

if (requireNamespace("ggplot2", quietly = TRUE)) {
  td <- tempdir()
  save_plot(plot_ftir(biodiesel), filename = file.path(td, "biodiesel_plot.png"))
}
#> Saving 6.67 x 6.67 in image