Shared parameters

These parameters are used in multiple functions. They are defined here to make them easier to import and to find.

Arguments

call

(`environment`) The caller environment for error messages. L'environnement de l'appelant pour les messages d'erreur.

ftir

A data.frame in long format with columns `sample_id`, `wavenumber`, and `absorbance`/`transmittance`. Un data.frame au format long avec les colonnes `sample_id`, `wavenumber`, et `absorbance`/`transmittance`.

ftir_spectra_plot

A plot generated by [plot_ftir()] or [plot_ftir_stacked()]. Un tracé généré par [plot_ftir()] ou [plot_ftir_stacked()].

lang

Optional language argument: `fr`/`french`/`francais`/`français` for French; `en`/`english`/`anglais` for English; `NA` uses default. Argument optionnel pour la langue : `fr`/`french`/`francais`/`français` pour le français ; `en`/`english`/`anglais` pour l'anglais ; `NA` utilise la valeur par défaut.

label_aesthetics

A named `list` of aesthetics to pass to ggplot for creating the label. See `[ggplot2::geom_text()]`'s aesthetics section for more info. Specifically, `alpha`, `colo(u)r`, `family`, `fill`, `fontface` and `size`are permitted. Positioning aesthetics will be removed.

Une `list` nommée d'esthétiques à transmettre à ggplot pour créer l'étiquette. Voir la section esthétique de `[ggplot2::geom_text()]` pour plus d'informations. Plus précisément, `alpha`, `colo(u)r`, `family`, `fill`, `fontface` et `size` sont autorisés. Les aspects esthétiques du positionnement seront supprimés.

plot_title

Title for plot. Defaults to "FTIR Spectra". Vector length 2 uses second element as subtitle. Titre du tracé. Par défaut "FTIR Spectra". Un vecteur de longueur 2 utilise le deuxième élément comme sous-titre.

legend_title

Title for legend. Defaults to "Sample ID". Titre de la légende. Par défaut "Sample ID".